style: DRY tanzil city component
